fixed functionality and errors in resetPassword property

+++ aplicacion_backend/turisteczbackend/src/main/java/org/turistecz/turisteczbackend/repository/VerificationTokenRepository.java
@@ -8,7 +8,7 @@ import org.turistecz.turisteczbackend.model.VerificationToken;
 import java.util.Optional;
 
 @Repository
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Integer> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
 
     Optional<VerificationToken> findByToken(String token);
     Optional<VerificationToken> findByTokenAndTipo(String token, String tipo);
